+++ app/assets/locales/messages_ccodk_default.txt
@@ -610,6 +610,9 @@ wifi.direct.connect.failed=Connecting failed, please retry
 wifi.direct.error.wiping.forms=Error wiping forms: ${0}
 wifi.direct.connect.success=Successfully connected to peer
 
+wifi.direct.pull.successful=Record pull successful, attempting to zip files...
+wifi.direct.pull.unsuccessful=Pulling records from storage failed.
+
 wifi.direct.zip.task.title=Zipping Forms...
 wifi.direct.zip.task.message=CommCare is compressing your data for transfer
 wifi.direct.unzip.task.title=Unzipping Forms...
@@ -622,9 +625,11 @@ wifi.direct.receive.task.title=Started Receiving Files
 wifi.direct.receive.task.message=CommCare is receiving files
 wifi.direct.wipe.task.title=Wiping Forms
 wifi.direct.wipe.task.message=Cleaning up after transfer
+wifi.direct.pull.task.title=Pulling Forms
+wifi.direct.pull.task.message=Pulling forms from record storage.
 
 wifi.direct.status.transfer.header=You are in Transfer Form Mode
-wifi.direct.status.transfer.count=Device has ${0} unsent forms.
+wifi.direct.status.transfer.count=Device has ${0} local forms and ${1} received forms.
 wifi.direct.status.transfer.message=This will allow you to transfer forms from this device to another device via Wi-Fi Direct
 
 wifi.direct.status.receive.header=You are in Receive Form Mode
